Removed output start marker
- Removed insert_output_start_marker() as only end marker needed for
  read() in parse_debugger_output()
- Moved insert_output_end_marker() to utilities.c
