+++ src/start_debugger.c
@@ -12,7 +12,6 @@
 #include "parse_debugger_output.h"
 #include "utilities.h"
 #include "plugins.h"
-#include "insert_output_marker.h"
 #include "update_window_data/_update_window_data.h"
 #include "update_window_data/get_binary_path_time.h"
